pyproject.toml
[project]
name = "tf-mac-env"
version = "0.1.0"
description = "TensorFlow environment for Apple Silicon"
readme = "README.md"
# tensorflow-metalとの互換性を考慮し、Python 3.10/3.11系を推奨
requires-python = ">=3.10, <3.12"
dependencies = [
    "tensorflow-macos==2.15.0",
    "tensorflow-metal==1.1.0",
    "ipykernel"
]